requirements.txt
requests~=2.32.3
python-decouple~=3.8
pandas~=2.2.2
dash~=2.17.1
dash-bootstrap-components~=1.6.0
Flask-Caching==2.3.0
gunicorn~=23.0.0